clients/pledge: HowItWorks standalone component
